% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenity.R
\docType{package}
\name{serenity}
\alias{serenity}
\alias{serenity-package}
\alias{serenity}
\title{Serenity: Data science in the classroom.}
\usage{
serenity()
}
\description{
Serenity: Data science in the classroom.

Launch Serenity in the default browser.
}
\examples{
\dontrun{
serenity::serenity()
}
}
